+++ playground/app/src/app/Library/TaskServices/Projector.ts
