--- playground/app/src/app/Modules/Todo/Picker/Component.ts
@@ -1,94 +1,0 @@
-import { CdkDragDrop } from "@angular/cdk/drag-drop";
-import { Component, OnDestroy, OnInit } from "@angular/core";
-import { ActivatedRoute } from "@angular/router";
-import { DOCUMENT_TITLE, LedgerService, ParamsService, TitleService } from "@valkyr/angular";
-import { ModalService } from "@valkyr/angular/src/Components/Modal/Service";
-import { IdentityService } from "@valkyr/identity";
-import { WorkspaceStore } from "stores";
-
-import { WorkspaceService } from "../../../Library/WorkspaceServices";
-import { CreateTodoDialog } from "../Dialogues/CreateTodo/Component";
-import { Todo } from "../Models/Todo";
-import { TodoService } from "../Services/Todo";
-
-@Component({
-  selector: "todo-picker",
-  templateUrl: "./Template.html"
-})
-export class TodoPickerComponent implements OnInit, OnDestroy {
-  todos: Todo[] = [];
-
-  name = "";
-
-  constructor(
-    readonly workspace: WorkspaceService,
-    readonly todo: TodoService,
-    readonly ledger: LedgerService,
-    readonly modal: ModalService,
-    readonly title: TitleService,
-    readonly params: ParamsService,
-    readonly route: ActivatedRoute,
-    readonly identity: IdentityService
-  ) {
-    title.set("Todos", DOCUMENT_TITLE, "workspace");
-  }
-
-  ngOnInit(): void {
-    const workspaceId = this.workspace.selected;
-    if (!workspaceId) {
-      throw new Error("TodoPickerComponent Violation: Could not resolve current workspace");
-    }
-    this.#loadWorkspace(workspaceId);
-    this.#loadTodos(workspaceId);
-  }
-
-  ngOnDestroy(): void {
-    this.workspace.unsubscribe(this);
-    this.todo.unsubscribe(this);
-  }
-
-  #loadWorkspace(workspaceId: string) {
-    this.workspace.subscribe(this, { criteria: { id: workspaceId }, limit: 1 }, (workspace) => {
-      if (workspace) {
-        this.title.set(`${workspace.name} Todos`, DOCUMENT_TITLE, "workspace");
-      }
-    });
-  }
-
-  #loadTodos(workspaceId: string) {
-    this.todo.subscribe(
-      this,
-      {
-        criteria: { workspaceId },
-        stream: {
-          aggregate: "todo",
-          endpoint: `/workspaces/${workspaceId}/todos`
-        }
-      },
-      (todos) => {
-        this.todos = todos;
-      }
-    );
-  }
-
-  public openAddTodo() {
-    this.modal.open(CreateTodoDialog);
-  }
-
-  public async drop(event: CdkDragDrop<string[]>) {
-    if (event.previousContainer === event.container) {
-      if (!this.workspace.selected) {
-        throw new Error("Could not resolve workspace id");
-      }
-      const workspace = await this.ledger.reduce(this.workspace.selected, WorkspaceStore.Workspace);
-      if (!workspace) {
-        throw new Error("Could not resolve workspace");
-      }
-      const member = workspace.members.getById(this.identity.auditor);
-      if (!member) {
-        throw new Error("Could not resolve workspace member");
-      }
-      this.todo.move(event.item.data.id, event.currentIndex, member.id);
-    }
-  }
-}
